refactor and rename directories
