+++ MANIFEST.in
@@ -1,5 +1,5 @@
 include LICENSE
 include README.md
-recursive-include sensorAtlas/routers *
+recursive-include sensoratlas/routers *
 recursive-include tests *
 recursive-include docs *
