+++ sensoratlas/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class sensorAtlasConfig(AppConfig):
-    name = 'sensorAtlas'
+    name = 'sensoratlas'
     verbose_name = 'Sensor Atlas'
