+++ sensoratlas/errors.py
