+++ sensoratlas/utils.py
