+++ tests/test_complianceA1.py
@@ -1,7 +1,7 @@
 from rest_framework.reverse import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
-from sensorAtlas.models import Thing, Location, HistoricalLocation, \
+from sensoratlas.models import Thing, Location, HistoricalLocation, \
     Datastream, Sensor, ObservedProperty, Observation, FeatureOfInterest
 from django.contrib.gis.geos import Point, Polygon
 from django.utils import timezone
